Moved configuration related code into service, updated the tests, and have the injection module handling the configuration now.
