+++ src/main/java/liuni/dagger/ServiceInjectionModule.java
@@ -2,6 +2,7 @@ package liuni.dagger;
 
 import dagger.Module;
 import dagger.Provides;
+import liuni.configs.LiUniConfig;
 import liuni.services.TwitterService;
 
 import javax.inject.Singleton;
@@ -9,9 +10,10 @@ import javax.inject.Singleton;
 @Module
 public class ServiceInjectionModule {
     private static TwitterService twitterService;
+    LiUniConfig config;
 
-    public ServiceInjectionModule() {
-
+    public ServiceInjectionModule(LiUniConfig config) {
+        this.config = config;
     }
 
     @Provides
@@ -21,6 +23,7 @@ public class ServiceInjectionModule {
             return twitterService;
         }
         twitterService = TwitterService.getInstance();
+        twitterService.setUpConfiguration(config.getTwitter());
         return twitterService;
     }
 }
