+++ src/test/java/liuni/LiUniResourceTest.java
@@ -19,12 +19,7 @@ import java.util.List;
 import java.util.Optional;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotEquals;
-import static org.mockito.ArgumentMatchers.isA;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 public class LiUniResourceTest {
@@ -40,7 +35,6 @@ public class LiUniResourceTest {
         when(mockedConfig.getTwitterAccounts()).thenReturn(new ArrayList<TwitterAccountConfig>());
 
         resource = new LiUniResource(mockedService);
-        resource.setUpConfiguration(mockedConfig);
     }
 
     @Test
@@ -210,121 +204,6 @@ public class LiUniResourceTest {
         catch (Exception e) {
             Assert.fail("This exception is not expected.");
         }
-    }
-
-    @Test
-    public void testGetSetConfig() {
-        TwitterConfig retrievedConfig = resource.getConfig();
-        TwitterConfig mockedTwitterConfig = mock(TwitterConfig.class);
-        resource.setConfig(mockedTwitterConfig);
-        assertNotEquals(retrievedConfig, mockedTwitterConfig);
-    }
-
-    @Test
-    public void testSetConfigIndex() {
-        int newIndex = 1;
-
-        assertNotEquals(newIndex, resource.getDefaultAccountIndex());
-
-        List<TwitterAccountConfig> accConfigList = new ArrayList<TwitterAccountConfig>();
-        accConfigList.add(mock(TwitterAccountConfig.class));
-        accConfigList.add(mock(TwitterAccountConfig.class));
-        when(mockedConfig.getTwitterAccounts()).thenReturn(accConfigList);
-
-        resource.setConfig(mockedConfig);
-        resource.setConfigIndex(newIndex);
-
-        assertEquals(newIndex, resource.getDefaultAccountIndex());
-    }
-
-    @Test
-    public void testSetConfigIndexOutofBoundsHigh() {
-        int newIndex = 1;
-
-        assertNotEquals(newIndex, resource.getDefaultAccountIndex());
-
-        List<TwitterAccountConfig> accConfigList = new ArrayList<TwitterAccountConfig>();
-        accConfigList.add(mock(TwitterAccountConfig.class));
-        when(mockedConfig.getTwitterAccounts()).thenReturn(accConfigList);
-
-        resource.setConfig(mockedConfig);
-        resource.setConfigIndex(newIndex);
-
-        assertNotEquals(newIndex, resource.getDefaultAccountIndex());
-    }
-
-    @Test
-    public void testSetConfigIndexOutofBoundLow() {
-        int newIndex = -1;
-
-        assertNotEquals(newIndex, resource.getDefaultAccountIndex());
-
-        List<TwitterAccountConfig> accConfigList = new ArrayList<TwitterAccountConfig>();
-        accConfigList.add(mock(TwitterAccountConfig.class));
-        when(mockedConfig.getTwitterAccounts()).thenReturn(accConfigList);
-
-        resource.setConfig(mockedConfig);
-        resource.setConfigIndex(newIndex);
-
-        assertNotEquals(newIndex, resource.getDefaultAccountIndex());
-    }
-
-    @Test
-    public void testConfigNull() {
-        resource = new LiUniResource(mockedService);
-        resource.setUpConfiguration(null);
-        assertEquals(null, resource.getConfig());
-        assertEquals(mockedService, resource.getTwitterService());
-    }
-
-    //Review tests from here to below, are they still valid?
-    @Test
-    public void testConfigListNotEmptyIndexHigh() {
-        List<TwitterAccountConfig> mockedAccConfigs = new ArrayList<TwitterAccountConfig>();
-        mockedAccConfigs.add(mock(TwitterAccountConfig.class));
-        mockedAccConfigs.add(mock(TwitterAccountConfig.class));
-        when(mockedConfig.getTwitterAccounts()).thenReturn(mockedAccConfigs);
-        when(mockedConfig.getDefaultAccountIndex()).thenReturn(mockedAccConfigs.size()+1);
-
-        resource = new LiUniResource(mockedService);
-        resource.setUpConfiguration(mockedConfig);
-
-        assert(resource.getDefaultAccountIndex() >= mockedAccConfigs.size());
-        assertEquals(mockedService, resource.getTwitterService());
-        verify(mockedService, never()).setTwitterAccountConfig(isA(TwitterAccountConfig.class));
-    }
-
-    @Test
-    public void testConfigListNotEmptyIndexLow() {
-        List<TwitterAccountConfig> mockedAccConfigs = new ArrayList<TwitterAccountConfig>();
-        mockedAccConfigs.add(mock(TwitterAccountConfig.class));
-        mockedAccConfigs.add(mock(TwitterAccountConfig.class));
-        when(mockedConfig.getTwitterAccounts()).thenReturn(mockedAccConfigs);
-        when(mockedConfig.getDefaultAccountIndex()).thenReturn(-1);
-
-        resource = new LiUniResource(mockedService);
-        resource.setUpConfiguration(mockedConfig);
-
-        assert(resource.getDefaultAccountIndex() < 0);
-        assertEquals(mockedService, resource.getTwitterService());
-        verify(mockedService, never()).setTwitterAccountConfig(isA(TwitterAccountConfig.class));
-    }
-
-    @Test
-    public void testConfigListNotEmptyIndexAppropriate() {
-        List<TwitterAccountConfig> mockedAccConfigs = new ArrayList<TwitterAccountConfig>();
-        mockedAccConfigs.add(mock(TwitterAccountConfig.class));
-        mockedAccConfigs.add(mock(TwitterAccountConfig.class));
-        when(mockedConfig.getTwitterAccounts()).thenReturn(mockedAccConfigs);
-        when(mockedConfig.getDefaultAccountIndex()).thenReturn(0);
-
-        resource = new LiUniResource(mockedService);
-        resource.setUpConfiguration(mockedConfig);
-
-        assert(resource.getDefaultAccountIndex() >= 0);
-        assert(resource.getDefaultAccountIndex() < mockedAccConfigs.size());
-        assertEquals(mockedService, resource.getTwitterService());
-        verify(mockedService, times(1)).setTwitterAccountConfig(isA(TwitterAccountConfig.class));
     }
 
 }
